+++ tests/roots/test-ext-collapse/conf.py
@@ -1,0 +1,3 @@
+project = 'test-directive-only'
+exclude_patterns = ['_build']
+extensions = ['sphinx.ext.collapse']