+++ CHANGES.rst
@@ -21,6 +21,9 @@ Features added
 * Flatten ``Union[Literal[T], Literal[U], ...]`` to ``Literal[T, U, ...]``
   when turning annotations into strings.
   Patch by Adam Turner.
+* #10532: Add a new extension to support collapsible content in HTML,
+  ``sphinx.ext.collapse``, which enables the :rst:dir:`collapse` directive.
+  Patch by Adam Turner.
 
 Bugs fixed
 ----------